entity 양방향 참조 하게 수정 - 정리 후 재수정 필요

+++ src/main/kotlin/com/note/resource/repository/postAttachmentFile/PostAttachmentFileRepository.kt
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository
 
 @Repository
 interface PostAttachmentFileRepository: JpaRepository<PostAttachmentFile, Number>{
-    fun findBySeqIdAndDeleteFlagIs(seqId: Int, deleteFlag: Boolean = false): PostAttachmentFile
+    fun findBySeqIdAndDeleteFlagIs(seqId: Long, deleteFlag: Boolean = false): PostAttachmentFile
 }
